Windaube doesn't know ssize_t, hence using int.

+++ libqimessaging/qimessaging/buffer.hpp
@@ -22,8 +22,8 @@ namespace qi
   public:
     Buffer();
 
-    ssize_t write(const void *data, size_t size);
-    ssize_t read(void *data, size_t size);
+    int    write(const void *data, size_t size);
+    int    read(void *data, size_t size);
     // equivalent to peek() && seek()
     void  *read(size_t size);
     size_t size() const;
